pyproject.toml
[tool.poetry]
name = "clueless"
version = "0.1.0"
description = ""
authors = ["tzeng7 <thebosstzeng@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
PodSixNet = "^0.11.0"
pygame-gui = "^0.6.9"
pygame-ce = "^2.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
